refactor: renamed some manifest types for clarity
