+++ crates/pixi_manifest/src/manifests/table_name.rs
